# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Currency Exposure Calculator API
  version: "1.0.0"
  description: |
    Calculates real-time currency exposure for authorized-but-not-captured transactions.
    Rates are fetched from a live external FX API.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
                example: ok
  /v1/exposure/calculate:
    post:
      summary: Calculate per-transaction and aggregate FX exposure
      operationId: calculateExposure
      description: |
        Primary production endpoint. Expects explicit transactions in the request body.
        Empty payloads return `NO_TRANSACTIONS`.
      parameters:
        - in: header
          name: X-API-Key
          required: true
          schema:
            type: string
          description: API authentication key.
        - in: header
          name: X-Idempotency-Key
          required: true
          schema:
            type: string
            format: uuid
          description: Idempotency key used for 24-hour replay semantics on POST requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateExposureInputRequest"
            examples:
              standard:
                value:
                  risk_threshold_percentage: 2
                  transactions:
                    - transaction_id: TXN-1001
                      authorization_timestamp: "2026-06-20T10:30:00Z"
                      authorized_amount: 14400
                      presentment_currency: BRL
                      settlement_currency: EUR
                      authorization_rate: 6.0
      responses:
        "200":
          description: Exposure successfully calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculateExposureResponse"
        "400":
          description: Invalid request payload, empty transaction list, or unsupported input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Live rate unavailable from upstream API
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/exposure/calculate/test:
    post:
      summary: "[DEMO-ONLY] Calculate exposure using configured analytics test dataset"
      operationId: calculateExposureTest
      description: |
        Demo-only endpoint for analytics and QA workflows.
        It always loads transactions from the configured test-data file at call time.
      parameters:
        - in: header
          name: X-API-Key
          required: true
          schema:
            type: string
          description: API authentication key.
        - in: header
          name: X-Idempotency-Key
          required: true
          schema:
            type: string
            format: uuid
          description: Idempotency key used for 24-hour replay semantics on POST requests.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateExposureTestRequest"
      responses:
        "200":
          description: Exposure successfully calculated from default test data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculateExposureResponse"
        "400":
          description: Invalid request payload or unsupported input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Live rate unavailable from upstream API
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    CalculateExposureInputRequest:
      type: object
      required:
        - transactions
      properties:
        risk_threshold_percentage:
          type: number
          format: double
          minimum: 0
          default: 2
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/PendingTransaction"
    CalculateExposureTestRequest:
      type: object
      properties:
        risk_threshold_percentage:
          type: number
          format: double
          minimum: 0
          default: 2
    PendingTransaction:
      type: object
      required:
        - account_id
        - payment_id
        - transaction_id
        - merchant_order_id
        - country
        - provider
        - payment_method_type
        - payment_status
        - transaction_status
        - authorized_at
        - authorization_expires_at
        - authorized_amount
        - capture_amount
        - presentment_currency
        - settlement_currency
      properties:
        account_id:
          type: string
        payment_id:
          type: string
        transaction_id:
          type: string
        merchant_order_id:
          type: string
        country:
          type: string
        provider:
          type: string
        payment_method_type:
          type: string
        payment_status:
          type: string
        transaction_status:
          type: string
        authorization_timestamp:
          type: string
          format: date-time
        authorized_at:
          type: string
          format: date-time
        authorization_expires_at:
          type: string
          format: date-time
        authorized_amount:
          type: number
          format: double
          minimum: 0
        capture_amount:
          type: number
          format: double
          minimum: 0
        presentment_currency:
          type: string
          minLength: 3
          maxLength: 3
        settlement_currency:
          type: string
          minLength: 3
          maxLength: 3
        authorization_rate:
          type: number
          format: double
          minimum: 0
          nullable: true
          description: |
            Presentment-per-settlement rate at authorization time.
            When omitted or zero, the service attempts historical-rate lookup.
        authorization_rate_source:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    CalculateExposureResponse:
      type: object
      properties:
        summary:
          $ref: "#/components/schemas/ExposureSummary"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionExposure"
        ranking:
          type: array
          items:
            $ref: "#/components/schemas/RiskRankingItem"
    ExposureSummary:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        total_exposure_amount:
          type: number
          format: double
        gain_count:
          type: integer
        loss_count:
          type: integer
        neutral_count:
          type: integer
        high_risk_count:
          type: integer
        high_risk_exposure_total:
          type: number
          format: double
        stale_rate_transaction_count:
          type: integer
        currency_breakdown:
          type: array
          items:
            $ref: "#/components/schemas/CurrencyExposureBreakdown"
        risky_currency_pairs:
          type: array
          items:
            $ref: "#/components/schemas/CurrencyPairRiskInsight"
        best_transaction:
          $ref: "#/components/schemas/TransactionExposure"
        worst_transaction:
          $ref: "#/components/schemas/TransactionExposure"
    CurrencyExposureBreakdown:
      type: object
      properties:
        presentment_currency:
          type: string
        total_exposure_amount:
          type: number
          format: double
        transaction_count:
          type: integer
        gain_count:
          type: integer
        loss_count:
          type: integer
    CurrencyPairRiskInsight:
      type: object
      properties:
        presentment_currency:
          type: string
        settlement_currency:
          type: string
        transaction_count:
          type: integer
        high_risk_count:
          type: integer
        total_exposure_amount:
          type: number
          format: double
        average_exposure_percentage:
          type: number
          format: double
        trend:
          type: string
          enum: [dangerous, stable_or_favorable]
    TransactionExposure:
      type: object
      properties:
        account_id:
          type: string
        payment_id:
          type: string
        merchant_order_id:
          type: string
        transaction_id:
          type: string
        authorization_timestamp:
          type: string
          format: date-time
        authorization_expires_at:
          type: string
          format: date-time
        presentment_currency:
          type: string
        settlement_currency:
          type: string
        authorized_amount:
          type: number
          format: double
        capture_amount:
          type: number
          format: double
        original_settlement_amount:
          type: number
          format: double
        current_settlement_amount:
          type: number
          format: double
        exposure_amount:
          type: number
          format: double
        exposure_percentage:
          type: number
          format: double
        authorization_rate:
          type: number
          format: double
        authorization_rate_source:
          type: string
        current_rate:
          type: number
          format: double
        current_rate_timestamp:
          type: string
          format: date-time
        quote_age_seconds:
          type: number
          format: double
        current_rate_source:
          type: string
        is_stale_rate:
          type: boolean
        is_high_risk:
          type: boolean
        eligible_to_capture:
          type: boolean
        capture_endpoint_hint:
          type: string
        capture_type:
          type: string
          enum: [full, partial]
        partial_capture_supported:
          type: boolean
        authorization_expiry_risk:
          type: string
          enum: [expired, high, medium, low, unknown]
        blocking_reason:
          type: string
        expected_loss_avoided:
          type: number
          format: double
        urgency_score:
          type: number
          format: double
        expiry_score:
          type: number
          format: double
        fx_severity_score:
          type: number
          format: double
        next_action:
          type: string
        recommendation:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    RiskRankingItem:
      type: object
      properties:
        rank:
          type: integer
        transaction_id:
          type: string
        presentment_currency:
          type: string
        settlement_currency:
          type: string
        exposure_amount:
          type: number
          format: double
        exposure_percentage:
          type: number
          format: double
        is_high_risk:
          type: boolean
        recommendation:
          type: string
          enum: [capture_now, monitor]
    ErrorResponse:
      type: object
      required:
        - type
        - code
        - message
      properties:
        type:
          type: string
          example: validation_error
        code:
          type: string
          example: INVALID_REQUEST_BODY
        message:
          type: string
          example: request body must be a valid JSON object
        details:
          type: object
          additionalProperties: true